requirements.txt
albucore==0.0.12
albumentations==1.4.11
efficientnet-pytorch==0.7.1
importlib-metadata==7.0.1
importlib_resources==6.1.2
Jinja2==3.1.3
joblib==1.3.2
lazy_loader==0.3
matplotlib==3.7.5
numpy==1.24.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.10.0.84
pandas==2.0.3
pillow==10.2.0
pip==23.3.1
pretrainedmodels==0.7.4
requests==2.31.0
requests-oauthlib==1.3.1
safetensors==0.4.2
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
segmentation-models-pytorch==0.3.3
timm==0.9.2
torch==2.2.1
torchaudio==2.2.1
torchvision==0.17.1
tqdm==4.66.4
urllib3==2.1.0